[Refactor] #3 SpanText 분리
