+++ app/src/main/java/com/sopt/yeogieottae/ui/room/RoomFragment.kt
@@ -1,19 +1,13 @@
 package com.sopt.yeogieottae.ui.room
 
 import android.os.Bundle
-import android.text.Spannable
-import android.text.SpannableString
-import android.text.SpannableStringBuilder
-import android.text.Spanned
-import android.text.style.BulletSpan
-import android.text.style.ForegroundColorSpan
-import android.text.style.TextAppearanceSpan
 import android.view.View
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.viewModels
 import com.sopt.yeogieottae.R
 import com.sopt.yeogieottae.databinding.FragmentRoomBinding
 import com.sopt.yeogieottae.util.BaseFragment
+import com.sopt.yeogieottae.util.bulletSpanText
+import com.sopt.yeogieottae.util.bulletSpanTextCancel
 
 
 class RoomFragment : BaseFragment<FragmentRoomBinding>(
@@ -29,7 +23,8 @@ class RoomFragment : BaseFragment<FragmentRoomBinding>(
     private fun initView() {
         with(binding) {
             tvRoomDetailRoomName.text = roomViewModel.Room.value?.room_name
-            tvRoomSummaryDiscountPrice.text =  getString(R.string.all_price).format(roomViewModel.Room.value?.price)
+            tvRoomSummaryDiscountPrice.text =
+                getString(R.string.all_price).format(roomViewModel.Room.value?.price)
             tvRoomDetailBottomPrice.text =
                 getString(R.string.all_price).format(roomViewModel.Room.value?.price)
             tvRoomSummaryTime.text = getString(R.string.room_in_out).format(
@@ -37,88 +32,20 @@ class RoomFragment : BaseFragment<FragmentRoomBinding>(
                 roomViewModel.Room.value?.end_time
             )
             tvRoomDetailNormalInfoDescription.text =
-                createSpannedText(roomViewModel.detail.value?.information ?: emptyList())
+                bulletSpanText(
+                    requireContext(),
+                    roomViewModel.detail.value?.information ?: emptyList()
+                )
             tvRoomDetailFacilitiesDescription.text =
-                createSpannedText(roomViewModel.detail.value?.facilities ?: emptyList())
+                bulletSpanText(
+                    requireContext(),
+                    roomViewModel.detail.value?.facilities ?: emptyList()
+                )
             tvRoomDetailCancelDescription.text =
-                createSpannedTextCancel(roomViewModel.detail.value?.cancel ?: emptyList())
-        }
-    }
-
-    private fun createSpannedText(
-        texts: List<String>,
-        bulletColor: Int = ContextCompat.getColor(requireContext(), R.color.gray_700)
-    ): SpannableStringBuilder {
-        val spannableStringBuilder = SpannableStringBuilder()
-
-        for ((index, text) in texts.withIndex()) {
-            val spannableString = SpannableString(text)
-            val bulletSpan = BulletSpan(3, bulletColor, 1)
-
-            spannableString.setSpan(
-                bulletSpan,
-                0,
-                spannableString.length,
-                Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
-            )
-            spannableStringBuilder.append(spannableString)
-            if (index < texts.size - 1) {
-                spannableStringBuilder.append("\n")
-            }
-        }
-        return spannableStringBuilder
-    }
-
-    private fun createSpannedTextCancel(
-        texts: List<String>,
-        bulletColor: Int = ContextCompat.getColor(requireContext(), R.color.gray_700)
-    ): SpannableStringBuilder {
-        val spannableStringBuilder = SpannableStringBuilder()
-        for ((index, text) in texts.withIndex()) {
-            val spannableString = SpannableString(text)
-            val bulletSpan = BulletSpan(3, bulletColor, 1)
-
-            spannableString.setSpan(
-                bulletSpan,
-                0,
-                spannableString.length,
-                Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
-            )
-            if (index == 3) {
-                spannableString.setSpan(
-                    ForegroundColorSpan(ContextCompat.getColor(requireContext(), R.color.cuation)),
-                    0,
-                    spannableString.length,
-                    Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
+                bulletSpanTextCancel(
+                    requireContext(),
+                    roomViewModel.detail.value?.cancel ?: emptyList()
                 )
-                spannableString.setSpan(
-                    TextAppearanceSpan(
-                        requireContext(),
-                        R.style.TextAppearance_Yeogieotte_Pretendard_h5
-                    ),
-                    0,
-                    spannableString.length,
-                    Spannable.SPAN_EXCLUSIVE_EXCLUSIVE
-                )
-            }
-
-            if (index in 4..5) {
-                spannableString.setSpan(
-                    TextAppearanceSpan(
-                        requireContext(),
-                        R.style.TextAppearance_Yeogieotte_Pretendard_h5
-                    ),
-                    0,
-                    spannableString.length,
-                    Spannable.SPAN_EXCLUSIVE_EXCLUSIVE
-                )
-            }
-            spannableStringBuilder.append(spannableString)
-
-            if (index < texts.size - 1) {
-                spannableStringBuilder.append("\n")
-            }
         }
-        return spannableStringBuilder
     }
 }
